I added the busiestHour, quiestestHour and twoHourBusiest
I used a while loop instead of for loop to figure out how to get busiest hour. I realized that after the loop would run my index would be off because the if statement so I added an else and just increased the index even if the parameters of the if else were met or not. I then included this same layout for the quietest hour just going down instead of up..... And finally i did the same thing for twoHourBusiest the only thing I changed was adding another index for it to look at and changing the print out.